+++ sdk/internal/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.2 (Unreleased)
 * Exports `RecordMode`, `PlaybackMode`, and `LiveMode` for determining test mode
+* Renames `StartRecording` to `Start` and `StopRecording` to `Stop`
 * When running in `LiveMode` no traffic will be routed to the proxy and the `StartRecording`/`StopRecording` methods are no-ops.
 * Adds the following sanitizers: `BodyKeySanitizer`, `BodyRegexSanitizer`, `ContinuationSanitizer`, `GeneralRegexSanitizer`, `HeaderRegexSanitizer`, `OAuthResponseSanitizer`, `RemoveHeaderSanitizer`, `URISanitizer`, `URISubscriptionIDSanitizer`
 * Adds `ResetSanitizer` for removing all sanitizers in an active session.